ci/pyproject.toml
[tool.poetry]
name = "ci"
version = "0.1.0"
description = ""
authors = ["Andrew Vaccaro <andrew.v@jarv.us>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["dotenv"], version = "^1.10.9"}
gitpython = "^3.1.31"
invoke = "^2.1.3"
doit = "^0.36.0"
pyyaml = "^6.0"
google-cloud-secret-manager = "^2.16.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
